+++ articular/__init__.py
@@ -1,2 +1,0 @@
-from articular.articular import Template
-from articular.document import ArticularResultDocument, ArticularSourceDocument