+++ krml/__init__.py
@@ -1,0 +1,2 @@
+from krml.krml import Template
+from krml.document import KrmlResultDocument, KrmlSourceDocument